* Define a base class `BaseNvcfSimulatorQPU` to encapsulate QPU subtype   that submits to NVCF and serves as parent class for C++ and Python classes.
* Docs fix
